+++ R/val_build.R
@@ -120,6 +120,8 @@ val_build <- function(
          call. = FALSE)
   }
   apply_verbose(verbose)
+  configure_bioc_repositories_if_requested(quiet = TRUE)
+  configure_riskmetric_offline_if_requested(quiet = TRUE)
 
   # Route pull_config() at any depth to the user-supplied config, if any.
   old_cfg <- options()["val.pipeline.config_path"]